fix: split up tests and refactored tests and implementation to promises
