+++ package.json
@@ -27,7 +27,7 @@
     "babelify": "^7.3.0",
     "browserify": "^13.1.0",
     "concurrently": "^2.2.0",
-    "eslint": "^3.2.2",
+    "eslint": "^3.5.0",
     "eslint-config-d-centralize": "^0.1.0",
     "jasmine-ajax": "^3.2.0",
     "jasmine-core": "^2.4.1",
